pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "research-navigator-agent"
version = "0.1.0"
description = "A LangGraph-based ReAct research agent with internal RAG and external web search"
authors = [
    {name = "Research Navigator Team"}
]
requires-python = ">=3.10"
dependencies = [
    "langchain>=0.3.0,<0.4.0",
    "langchain-community>=0.3.0,<0.4.0",
    "langchain-openai>=0.2.0,<0.3.0",
    "langgraph>=0.2.0,<0.3.0",
    "aiohttp>=3.8.3,<4.0.0",
    "faiss-cpu>=1.8.0,<2.0.0",
    "tavily-python>=0.5.0",
    "openai>=1.50.0",
    "httpx[socks]>=0.27.0",
    "python-dotenv>=1.0.0,<2.0.0",
    "pydantic>=2.0.0,<3.0.0",
    "pydantic-settings>=2.0.0,<3.0.0",
    "typer>=0.12.0,<1.0.0",
    "rich>=13.0.0,<14.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "black>=24.0.0",
    "ruff>=0.6.0",
]

[project.scripts]
research-nav = "src.cli.main:app"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.black]
line-length = 100
target-version = ['py310']

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
